Add smoke test
The smoke test runs the Java target application, attaches the profiler,
waits for 5 seconds, stops the profiler and dumps the results to a
file, and then parses the file looking for patterns that we know
must appear. This is a very basic smoke test that will only make sure
we haven't broken anything completely obvious.

The Makefile and Travis configuration are also updated to run the test.