Make perf and itimer use cpu clocks (#1)
* Default to itimer profiler when perf profiler isn't available

Today, it defaults to wall time profiler since
d78e0b3a517593ced9af0fbaa073105f379c228d. This commit doesn't make it
clear why that change of behavior, especially on Linux.

It is more useful to switch to itimer as it's behavior is closer to
perf sampling interval.

* Make perf and itimer actual oncpu profilers

Both itimer and perf would rely on wall clocks to sample threads. This
makes them wall-time profilers, not cpu profilers. To fix that, we
simply make sure that they both use cpu clocks.

The difference is for example in profiling `sleep 1`, a wall clock will
display 1 second of activity, while a cpu clock will display close to 0
seconds of activity. That's because, in the second case, the process was
sleeping for 1 second and was thus _not executing on cpu_ for 1 second.
The time spent executing on cpu was then close to 0.

* Add test for sleeping and CPU profiler