Support default pattern highlighting in flame graph
- Add an argument to flame graph args to pass in a regex pattern that
  will generate the graph with all matches highlighting in a different
color
- Primary use case would be to highlight user application stack lines
  from other java stack lines

It's often the case that seeing the difference between user code
  (often your application code is all under a shared package) and
non-user code quickly in a flame graph. The search function does let
users manually do this, however it's likely nice to be able to search
for other things while still having this basic distinction.

Also in a larger automated operation of generating flame graphs as part
of tooling, it can be nice to pre-search a relevant pattern for users,
such as having a presearch for application lines versus 3rd party java
lines as mentioned above.