Only use pinned memory for Kernel stack traces
This change removes the requirement for pinned memory when only
gathering user-mode stacktraces.

Pinned memory is a precious resource on some systems (some Linux
distributions only allow for a few hundred kilobytes) and changing the
limite requires admin privileges. This is particularely problematic when
running in a container without CAP_SYS_ADMIN.

In these limited cases, it's then better not to rely on pinned memory
when not required, and instead end up using perf only as a CPU-time
based trigger signaling mechanism.