Add argument to control live object sample buffer size
Previously, the buffer size was static at 1024.

As I've been experimenting with continuous live object sampling
and leak detection, I've found that I sometimes need to significantly
increase the size of this buffer in order to capture leaking objects.

Not all JVM workloads are equal and I'm not convinced a single static
value for the live object ring buffer capacity is right for all
workloads.

This commit adds an argument to control the size of the ring buffer.