Reduce mutex and rdlock iterations for nativelock tests
The commit includes reducing the mutex lock iteration count as in the
worst case, by the time the test canAsprofTraceAllLockTypes ends in 3
seconds, we may not have profiled rd and wr locks. Also reducing rd lock
iteration count for wr locks also to be sampled within the test
completion duration.